Allow SpinGroup tasks to report failure by returning special symbol
At the moment, the only way a SpinGroup task can report failure is by
raising an excepting. Since SpinGroup generates a backtrace from the
exception, task failures are very noisy.

This PR allows tasks to return a special symbol `:task_failed` to indicate
to the SpinGroup that it failed. During debrief, SpinGroup doesn't generate
a backtrace for these kind of failure.