Add a delay parameter to spinner
If specified, the SpinGroup will not print anything until it has been
running for at least that duration.

Not implemented:
If the task fails _before_ it has started printing the spinners,
it will debrief as usual, but without printing the failed spinner.
The task title is included in the debrief, so it doesn't seem like a big deal.