Clean up tests
I don't feel the need to refute the presence of the sentinel for jruby -
it seems more natural to just not run those tests for jruby since
they're meaningless when you can't `rescue Interrupt`.

For the other, rescuing Interrupt generated in code to `puts 'sentinel`
adequately asserts the intent of the test for mri/jruby.