Add spin group functionality
We allow turning the autodebrief off, but as it was, there was no way to
access the same task information the autodebrief used in order to
provide your own debrief. This addresses that. It also provides a method
to find out whether or not all tasks succeeded, which you can't
otherwise get with autodebrief.