Move handling the frame stack to thread variables
The frame stack used to be handled in environment variables, which
served as a global variable that we could serialize relevant information
we needed to track to. Now we want to support having frames handled
across threads (e.g. in SpinGroups) and so we store that information
in thread variables now.