Add support for interactive prompts in SpinGroup
SpinGroup runs Tasks inside of threads, which previously presented
difficulties for handling sources of input. In this change, we have
added support for a Mutex that supports reentrance as well as handling
global variables in thread variables. This way we can communicate a
thread's intent to read input, block other threads from doing so until
the input is collected, and move on safely.

Co-authored-by: Lisa Ugray <lisa.ugray@shopify.com>