Improve log output
This applies CLI::UI.enable_color? to Frame and Spinner output, and adds
CLI::UI.enable_cursor? to disable cursor control when it's not
supported.

I considered adding these as a parameters to Frame and Spinner, but it
ended up bloating the code for little tangible benefit as these settings
are usually static for the lifetime of the program. If we wanted to
support something like that, I think the best approach would be to
bundle color + cursor control settings along with IOLike as a sort of
printing context. Nothing here prevents us from doing that in the future
if the need arises.