Add SpinGroup.puts_above
This is useful for things like printing out a stream of logs while you
have one or more spinners reporting on the current status. We could also
add Spinner.puts_above, but when you have multiple Spinners, "above" is
kind of ambiguous. We can add it later if the need becomes obvious.